+++ CHANGELOG.md
@@ -35,6 +35,8 @@
   component will implement the `arez.component.Verifiable` interface and perform verification at runtime.
 * **\[processor\]** Update the compile time error generated when a component is empty so that it notes that if
   `@Dependency` is present then the component is considered non-empty.
+* **\[core\]** Add the ability to retrieve the priority of an observer or a computed using the spy subsytem via
+  either the `ObserverInfo.getPriority()` or the `ComputedValueInfo.getPriority()` methods.
 
 ### [v0.104](https://github.com/arez/arez/tree/v0.104) (2018-08-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.103...v0.104)